Don't build in alternative execroot
Initially this tool created a new symlink tree in a
`build/clangd_execroot` folder instead of placing an `external` symlink
directly in your source tree. It turns out that clangd mostly handles
this, but doesn't entirely. One example is that Ctrl-clicking (in
vscode) a function invocation would go to the declaration in a .h file
until you opened the .cc file, then it would click through to the
implementation.

All these problems go away when you get rid of the indirection and just
let clangd execute builds in your source root.